refactor(unit5): share the unified-host KV cache budget across backends
The default budget, the floor, the host reserve and the refuse-not-floor
arithmetic were private to the vLLM module. A second engine with the
same fill-the-pool default needs the identical policy, so it moves to
launch::admission and vLLM calls it.
