+++ src/backend/vllm/mod.rs
@@ -20,6 +20,10 @@ use super::{
 };
 use crate::daemon::context::MethodContext;
 use crate::daemon::probe::ProbeOptions;
+use crate::launch::admission::{
+  human_gib, unified_kv_cache_budget, DEFAULT_KV_CACHE_BYTES, MIN_KV_CACHE_BYTES,
+  UNIFIED_HOST_RESERVE_BYTES,
+};
 use crate::launch::params::LaunchParams;
 
 /// Stable backend id. The only place this string is authored.
@@ -416,7 +420,16 @@ impl Backend for VllmBackend {
     let cap = match snapshot.as_ref().filter(|_| sampled) {
       Some(s) => {
         let free = crate::launch::admission::effective_free_bytes(s);
-        match kv_cache_cap_bytes(free, weights_bytes) {
+        // On an APU, GPU memory *is* system RAM, and vLLM sizes its KV cache
+        // to fill whatever `gpu_memory_utilization` allows — a fraction of
+        // the **pool**, not of the model. Measured on a 121 GB Strix Halo:
+        // `0.15` on a 0.5B model reserved 15.1 GiB of KV cache (1.3M tokens,
+        // 644x concurrency for a 2048-token model) and cost 21.2 GB of RAM;
+        // the 0.92 default projects to ~106 GB and has frozen the machine
+        // outright. Clamping the *fraction* does not help, because the
+        // arithmetic is against the wrong number. Capping the cache in bytes
+        // does: vLLM then skips memory profiling entirely and honours it.
+        match unified_kv_cache_budget(free, weights_bytes) {
           Some(cap) => cap,
           None => {
             out.refusal = Some(format!(
@@ -424,10 +437,10 @@ impl Backend for VllmBackend {
                cache once the {} host reserve is kept free (host has {} \
                available). Lower --ctx, pick a smaller model, or set \
                kv_cache_memory_bytes to override.",
-              human_bytes(weights_bytes),
-              human_bytes(MIN_KV_CACHE_BYTES),
-              human_bytes(UNIFIED_HOST_RESERVE_BYTES),
-              human_bytes(free),
+              human_gib(weights_bytes),
+              human_gib(MIN_KV_CACHE_BYTES),
+              human_gib(UNIFIED_HOST_RESERVE_BYTES),
+              human_gib(free),
             ));
             return out;
           }
@@ -468,52 +481,6 @@ fn knob_f64(params: &LaunchParams, id: &str) -> Option<f64> {
 /// Whether the user pinned this knob (as opposed to leaving it to us).
 fn user_set(params: &LaunchParams, id: &str) -> bool {
   params.knobs.is_set_by_name_for(VLLM_BACKEND_ID, id)
-}
-
-/// Default KV cache budget when nothing else bounds it. Generous for a single
-/// user (~85x concurrency at 2k context on a 0.5B) and small enough that the
-/// launch cannot take the host down.
-const DEFAULT_KV_CACHE_BYTES: u64 = 8 * 1024 * 1024 * 1024;
-
-/// `1.5 GiB`-style label for a refusal message.
-fn human_bytes(b: u64) -> String {
-  const GIB: f64 = (1024 * 1024 * 1024) as f64;
-  format!("{:.1} GiB", b as f64 / GIB)
-}
-
-/// Floor for the cap. Below this the cache cannot serve a useful context, so
-/// the launch is refused outright rather than admitted with a token cache that
-/// would only fail after a full weight load.
-const MIN_KV_CACHE_BYTES: u64 = 512 * 1024 * 1024;
-
-/// Reserve left free for the OS and everything else after weights + cache.
-const UNIFIED_HOST_RESERVE_BYTES: u64 = 8 * 1024 * 1024 * 1024;
-
-/// The KV cache cap for a unified-memory host. **Always a value.**
-///
-/// On an APU, GPU memory *is* system RAM, and vLLM sizes its KV cache to fill
-/// whatever `gpu_memory_utilization` allows — a fraction of the **pool**, not
-/// of the model. Measured on a 121 GB Strix Halo: `0.15` on a 0.5B model
-/// reserved 15.1 GiB of KV cache (1.3M tokens, 644x concurrency for a
-/// 2048-token model) and cost 21.2 GB of RAM; the 0.92 default projects to
-/// ~106 GB and has frozen the machine outright. Clamping the *fraction* does
-/// not help, because the arithmetic is against the wrong number. Capping the
-/// cache in bytes does: vLLM then skips memory profiling entirely and honours
-/// the figure.
-///
-/// `None` means **refuse the launch**, not "no opinion".
-///
-/// An earlier version floored the cap instead of refusing, on the theory that
-/// the admission gate would catch the tight case. It cannot: the gate's demand
-/// is weights + *this* figure, so shrinking the figure shrinks the very term
-/// the gate evaluates, and a launch that should have been refused was admitted
-/// with the host reserve silently abandoned. Whoever decides there is not
-/// enough memory has to be whoever holds the number, which is here.
-fn kv_cache_cap_bytes(free_bytes: u64, weights_bytes: u64) -> Option<u64> {
-  let headroom = free_bytes
-    .saturating_sub(weights_bytes)
-    .saturating_sub(UNIFIED_HOST_RESERVE_BYTES);
-  (headroom >= MIN_KV_CACHE_BYTES).then(|| headroom.min(DEFAULT_KV_CACHE_BYTES))
 }
 
 impl VllmBackend {
@@ -916,28 +883,6 @@ mod tests {
       }
       other => panic!("expected a model-id poll, got {other:?}"),
     }
-  }
-
-  /// The freeze guard. vLLM's default sizes the KV cache against the pool,
-  /// which on a UMA host is system RAM — measured at ~106 GB of a 121 GB box.
-  #[test]
-  fn kv_cap_leaves_the_host_a_reserve_and_is_never_absent() {
-    const GB: u64 = 1024 * 1024 * 1024;
-    // Plenty free: the default budget applies, not "everything that fits".
-    assert_eq!(kv_cache_cap_bytes(113 * GB, GB), Some(8 * GB));
-    // Tight: the cap shrinks to what is left after weights + reserve.
-    assert_eq!(kv_cache_cap_bytes(20 * GB, 8 * GB), Some(4 * GB));
-    // Too tight to serve a useful context: refuse. Flooring instead used to
-    // shrink the demand the admission gate evaluates, so the gate could not
-    // fire and the launch went ahead with the host reserve abandoned.
-    assert_eq!(kv_cache_cap_bytes(10 * GB, 8 * GB), None);
-    assert_eq!(kv_cache_cap_bytes(4 * GB, 8 * GB), None);
-    assert_eq!(kv_cache_cap_bytes(0, 0), None);
-    // The exact boundary is admitted, not refused.
-    assert_eq!(
-      kv_cache_cap_bytes(8 * GB + MIN_KV_CACHE_BYTES, 0),
-      Some(MIN_KV_CACHE_BYTES)
-    );
   }
 
   #[test]
